+++ app/src/main/AndroidManifest.xml
@@ -13,17 +13,17 @@
         android:theme="@style/Theme.ATSOPTANDROID"
         tools:targetApi="31">
         <activity
-            android:name=".ui.theme.signup.SignUpPasswordActivity"
+            android:name=".presentation.signup.SignUpPasswordActivity"
             android:exported="false"
             android:label="@string/title_activity_sign_up_password"
             android:theme="@style/Theme.ATSOPTANDROID" />
         <activity
-            android:name=".ui.theme.signup.SignUpIdActivity"
+            android:name=".presentation.signup.SignUpIdActivity"
             android:exported="false"
             android:label="@string/title_activity_sign_up"
             android:theme="@style/Theme.ATSOPTANDROID" />
         <activity
-            android:name=".ui.theme.my.MyActivity"
+            android:name=".presentation.my.MyActivity"
             android:exported="false"
             android:label="@string/title_activity_my"
             android:theme="@style/Theme.ATSOPTANDROID" />
@@ -33,7 +33,7 @@
             android:label="@string/title_activity_sign_in"
             android:theme="@style/Theme.ATSOPTANDROID" />
         <activity
-            android:name=".ui.theme.signin.SignInActivity"
+            android:name=".presentation.signin.SignInActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.ATSOPTANDROID">
