+++ app/src/main/java/org/sopt/at/common/RippleEffect.kt
@@ -1,4 +1,4 @@
-package org.sopt.at.ui.common
+package org.sopt.at.common
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.interaction.MutableInteractionSource
