+++ app/src/main/java/org/sopt/at/ui/theme/Color.kt
@@ -1,7 +1,6 @@
 package org.sopt.at.ui.theme
 
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.GraphicsContext
 
 val Purple80 = Color(0xFFD0BCFF)
 val PurpleGrey80 = Color(0xFFCCC2DC)
